+++ frontend/src/components/Navigation/index.js
@@ -2,7 +2,8 @@ import React from 'react';
 import { NavLink } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 import ProfileButton from './ProfileButton';
-import LoginFormModal from '../LoginFormPage';
+import LoginFormModal from '../LoginFormModal';
+import SignupFormModal from '../SignupFormModal';
 import './Navigation.css';
 
 function Navigation({ isLoaded }) {
@@ -19,7 +20,8 @@ function Navigation({ isLoaded }) {
                 {/* <NavLink to="/login">Log In</NavLink> */}
                 <LoginFormModal />
                 <span> </span>
-                <NavLink to="/signup">Sign Up</NavLink>
+                <SignupFormModal />
+                {/* <NavLink to="/signup">Sign Up</NavLink> */}
             </>
         );
     }
