+++ frontend/src/components/SignupFormModal/index.js
@@ -1,0 +1,20 @@
+import React, { useState } from 'react';
+import { Modal } from '../../context/Modal';
+import SignupForm from './SignupForm';
+
+function SignupFormModal() {
+    const [showModal, setShowModal] = useState(false);
+
+    return (
+        <>
+            <button onClick={() => setShowModal(true)}>Log In</button>
+            {showModal && (
+                <Modal onClose={() => setShowModal(false)}>
+                    <SignupForm />
+                </Modal>
+            )}
+        </>
+    );
+}
+
+export default SignupFormModal;
